feat: Add MSB field for Horizontal Vane
Reading the horizontal vane will now ignore the MSB, which is returned
by some units for an unknown purpose. The remainder of the field is
still captured in order to better track possible unknown values.

The MSB is available at a new field specific for it.

At present, the MSB will be unset for SetSettings packets, as it does
not seem to be necessary. This may change in the future, but this is
at least a starting point.